Added table for mountable filters and lenses. Changed all id numbers from int to long. Replaced getLastObject methods with long returned from db.insert() method. Created methods in FilmDbHelper for building content values and reading objects from cursors.

+++ app/src/main/java/com/tommihirvonen/exifnotes/Activities/MainActivity.java
@@ -212,10 +212,10 @@ public class MainActivity extends AppCompatActivity implements RollsFragment.OnR
      * @param rollId The id of the roll that was pressed.
      */
     @Override
-    public void onRollSelected(int rollId){
+    public void onRollSelected(long rollId){
         FramesFragment newFragment = new FramesFragment();
         Bundle args = new Bundle();
-        args.putInt("ROLL_ID", rollId);
+        args.putLong("ROLL_ID", rollId);
         args.putBoolean("LOCATION_ENABLED", locationEnabled);
         newFragment.setArguments(args);
         getFragmentManager().beginTransaction()
